1.4.19

+++ ai-copilot-content-generator/classes/table.php
@@ -341,6 +341,10 @@ abstract class WaicTable {
 	 * @return query result
 	 */
 	public function delete( $where = '' ) {
+		if (empty($where)) {
+			return false;
+		}
+
 		$q = 'DELETE FROM ' . $this->_table;
 		if ($where) {
 			if (is_numeric($where)) {
